#1260 Move logic to modal
