+++ src/frontend/src/pages/FinancePage/ReimbursementRequestDetailPage/ReimbursementRequestDetailsView.tsx
@@ -3,27 +3,25 @@
  * See the LICENSE file in the repository root folder for details.
  */
 
-import { Grid, Typography, useTheme } from '@mui/material';
-import { Box } from '@mui/system';
-import { datePipe, fullNamePipe } from '../../../utils/pipes';
-import VerticalDetailDisplay from '../../../components/VerticalDetailDisplay';
 import { Edit } from '@mui/icons-material';
-import { useCurrentUser } from '../../../hooks/users.hooks';
-import { routes } from '../../../utils/routes';
-import ActionsMenu, { ButtonInfo } from '../../../components/ActionsMenu';
-import { useHistory } from 'react-router-dom';
-import PageLayout from '../../../components/PageLayout';
-import ReimbursementProductsView from './ReimbursementProductsView';
-import { ReimbursementRequest } from 'shared';
+import CheckIcon from '@mui/icons-material/Check';
+import ConfirmationNumberIcon from '@mui/icons-material/ConfirmationNumber';
 import DeleteIcon from '@mui/icons-material/Delete';
 import LocalShippingIcon from '@mui/icons-material/LocalShipping';
-import ConfirmationNumberIcon from '@mui/icons-material/ConfirmationNumber';
-import CheckIcon from '@mui/icons-material/Check';
-import { isReimbursementRequestApproved } from '../../../utils/reimbursement-request.utils';
-import AddSABONumberModal from './AddSABONumberModal';
+import { Grid, Typography, useTheme } from '@mui/material';
+import { Box } from '@mui/system';
 import { useState } from 'react';
-import { useSetSaboNumber } from '../../../hooks/finance.hooks';
-import { useToast } from '../../../hooks/toasts.hooks';
+import { useHistory } from 'react-router-dom';
+import { ReimbursementRequest } from 'shared';
+import ActionsMenu, { ButtonInfo } from '../../../components/ActionsMenu';
+import PageLayout from '../../../components/PageLayout';
+import VerticalDetailDisplay from '../../../components/VerticalDetailDisplay';
+import { useCurrentUser } from '../../../hooks/users.hooks';
+import { datePipe, fullNamePipe } from '../../../utils/pipes';
+import { isReimbursementRequestApproved } from '../../../utils/reimbursement-request.utils';
+import { routes } from '../../../utils/routes';
+import AddSABONumberModal from './AddSABONumberModal';
+import ReimbursementProductsView from './ReimbursementProductsView';
 
 interface ReimbursementRequestDetailsViewProps {
   reimbursementRequest: ReimbursementRequest;
@@ -31,23 +29,10 @@ interface ReimbursementRequestDetailsViewProps {
 
 const ReimbursementRequestDetailsView: React.FC<ReimbursementRequestDetailsViewProps> = ({ reimbursementRequest }) => {
   const theme = useTheme();
-  const toast = useToast();
   const totalCostBackgroundColor = theme.palette.mode === 'dark' ? theme.palette.grey[800] : theme.palette.grey[200];
   const user = useCurrentUser();
   const history = useHistory();
   const [addSaboNumberModalShow, setAddSaboNumberModalShow] = useState<boolean>(false);
-  const { mutateAsync: setSaboNumber } = useSetSaboNumber(reimbursementRequest.reimbursementRequestId);
-
-  const onSaboFormSubmit = async (data: { saboNumber: number }) => {
-    try {
-      await setSaboNumber(data);
-    } catch (error: unknown) {
-      if (error instanceof Error) {
-        toast.error(error.message);
-      }
-    }
-    setAddSaboNumberModalShow(false);
-  };
 
   const BasicInformationView = () => {
     return (
@@ -187,7 +172,7 @@ const ReimbursementRequestDetailsView: React.FC<ReimbursementRequestDetailsViewP
       <AddSABONumberModal
         modalShow={addSaboNumberModalShow}
         onHide={() => setAddSaboNumberModalShow(false)}
-        onSubmit={onSaboFormSubmit}
+        reimbursementRequestId={reimbursementRequest.reimbursementRequestId}
       />
     </PageLayout>
   );
